Group together all C-wrapping crates into a new c_wrappers directory
