+++ src/redisearch_rs/c_wrappers/buffer/src/writer.rs
