+++ src/redisearch_rs/c_wrappers/c_trie/src/lib.rs
