+++ src/redisearch_rs/c_wrappers/index_spec_cache/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "index_spec_cache"
+version.workspace = true
+edition.workspace = true
+license-file.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+ffi.workspace = true
+field_spec.workspace = true
+rm_array.workspace = true
+workspace_hack.workspace = true
