+++ src/redisearch_rs/c_wrappers/rm_array/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "rm_array"
+version.workspace = true
+edition.workspace = true
+license-file.workspace = true
+publish.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+ffi.workspace = true
+workspace_hack.workspace = true
