+++ src/redisearch_rs/headers/rlookup_rs.h
@@ -110,8 +110,6 @@ enum RLookup_Opt
 typedef uint32_t RLookup_Opt;
 #endif // __cplusplus
 
-typedef struct IndexSpecCache IndexSpecCache;
-
 /**
  * An append-only list of [`RLookupKey`]s.
  *
@@ -494,7 +492,7 @@ struct RLookup RLookup_New(void);
  * [valid]: https://doc.rust-lang.org/std/ptr/index.html#safety
  */
 void RLookup_SetCache(struct RLookup *lookup,
-                      struct IndexSpecCache *spcache);
+                      IndexSpecCache *spcache);
 
 /**
  * Returns `true` if this `RLookup` has an associated [`IndexSpecCache`].
