+++ src/redisearch_rs/rlookup/Cargo.toml
@@ -10,8 +10,12 @@ ffi.workspace = true
 c_ffi_utils.workspace = true
 document.workspace = true
 enumflags2.workspace = true
+field_spec.workspace = true
+index_spec.workspace = true
+index_spec_cache.workspace = true
 libc.workspace = true
 pin-project.workspace = true
+schema_rule.workspace = true
 sorting_vector.workspace = true
 strum.workspace = true
 value = { workspace = true, features = ["c_ffi_impl"] }
@@ -38,6 +42,7 @@ proptest = { workspace = true, features = ["std"] }
 value = { workspace = true }
 
 # Crate required to invoke C symbols
+field_spec = { workspace = true, features = ["unittest"] }
 rlookup = { path = ".", features = ["unittest"] }
 redisearch_rs = { path = "../c_entrypoint/redisearch_rs" }
 redis_mock.workspace = true
